refactor: introduce api typed facade
