+++ Cargo.toml
@@ -14,6 +14,7 @@ keywords = ["typescript", "javascript", "lint", "static-analysis", "code-quality
 categories = ["command-line-utilities", "development-tools"]
 
 [workspace.dependencies]
+fallow-api = { version = "2.102.0", path = "crates/api" }
 fallow-core = { version = "2.102.0", path = "crates/core" }
 fallow-config = { version = "2.102.0", path = "crates/config" }
 fallow-extract = { version = "2.102.0", path = "crates/extract" }
