+++ crates/cli/Cargo.toml
@@ -52,6 +52,7 @@ schema-emit = [
 ]
 
 [dependencies]
+fallow-api = { workspace = true }
 fallow-core = { workspace = true }
 fallow-config = { workspace = true }
 fallow-output = { workspace = true }
